backend/requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
python-multipart==0.0.20
opencv-python-headless==4.11.0.86
numpy==2.2.2
requests==2.32.3
scipy==1.15.1
pillow==11.1.0
python-dotenv==1.0.1
reportlab==4.2.5
